+++ src/components/Form/Select.tsx
@@ -24,7 +24,6 @@ const Select = ({
   required,
   enabled,
   SelectedOption,
-  defaultValue,
   startIcon,
   onIconClick,
   endIcons,
@@ -59,12 +58,11 @@ const Select = ({
         id={name}
         h={14}
         // value={defaultValue}
-        defaultValue={defaultValue}
 
         // pr={8}
       >
         {placeholder && (
-          <option value="" disabled={!enabled}>
+          <option value={""} hidden disabled={enabled}>
             {placeholder}
           </option>
         )}