+++ src/helper/checkTimeRange.ts
@@ -1,3 +1,5 @@
+import { differenceInYears } from "date-fns";
+
 export function isTimeInRange(
   start: string,
   end: string,
@@ -7,9 +9,6 @@ export function isTimeInRange(
   const endTime = getTimeInMinutes(end);
   const targetTime = getTimeInMinutes(target);
 
-  // if (startTime <= targetTime && targetTime < endTime) {
-  //   console.log(true);
-  // } else console.log(false);
   return startTime <= targetTime && targetTime < endTime;
 }
 
@@ -60,3 +59,13 @@ export function addOneHour(timeString: string): string {
 
   return `${updatedHours}:${updatedMinutes.toString().padStart(2, "0")}:00`;
 }
+
+export const calculateAge = (dob: Date): number => {
+  const age = differenceInYears(new Date(), dob);
+  return age;
+};
+
+export const getDayDifference = (date1: Date, date2: Date): number => {
+  const diffInMs = Math.abs(date2.getTime() - date1.getTime());
+  return Math.round(diffInMs / (1000 * 60 * 60 * 24));
+};