+++ src/service/service-api.ts
@@ -5,7 +5,7 @@ export const api = {
   signup: "user_management/otp-send/",
   otp_verify: "user_management/otp-verify/",
   register: "/doctor_consult/register/",
-  symptom: "/doctor_consult/symptom",
+  symptom: "core/symptoms/",
   academic: "/doctor_consult/academic-info/",
   certificate: "/doctor_consult/certificate-info/",
   experience: "/doctor_consult/experience-info/",
@@ -14,7 +14,7 @@ export const api = {
   experience_file: "doctor_consult/experience-files-upload/",
   doctor_profile: "/doctor_consult/doctor-profile/",
   edit_doctor_profile: "/doctor_consult/doctor-profile-update/",
-  specialization: "/core/specialization-list/",
+  specialization: "/core/specialization/",
   registereddoctor: "/dashboard/registered-doctor-list",
   pendingdoctor: "/dashboard/pending-doctor-list",
   approveddoctor: "/dashboard/approved-doctor-list",
@@ -30,6 +30,8 @@ export const api = {
   municipality: "/core/municipalities",
 
   doctor_availability: "/doctor_consult/doctor-availability/",
+  doctorProfileById: "/dashboard/view-doctor-profile/{id}",
+  rejectedDoctorList: "/dashboard/rejected-doctor-list/",
 };
 
 export interface NepMedsResponse<T = any> {